List of APIs defined in SimpleAiboWebAPI.

Setup APIs

SimpleAiboWebAPI.getDevices — Function

getDevices(accessToken)

tries to get device list associated to accessToken.

If successful,

returns device list of type Array{Dict}.

assigns accessToken to currentAccessToken, the local variable.

if non-empty device list is found, updates currentDevices and defaultDevice. Namely,

assigns found device list to currentDevices, the local variable.

assigns the first element of currentDevices to defaultDevice, the local variable, if currentDevices is not empty.

SimpleAiboWebAPI.getAccessToken — Function

getAccessToken()

returns the current access token, i.e., currentAccessToken, the local variable of this package.

Basic APIs

SimpleAiboWebAPI.getCurrentDevices — Function

getCurrentDevices()

returns the current device list, i.e., getCurrentDevices, the local variable of this package.

SimpleAiboWebAPI.getDefaultDevice — Function

getDefaultDevice()

returns the default device to which the WebAPI directs when target device is not specified in the call of askAction function.

SimpleAiboWebAPI.getDefaultDeviceId — Function

getDefaultDeviceId()

returns the deviceId of the default device.

SimpleAiboWebAPI.getDefaultDeviceNickname — Function

getDefaultDeviceNickname()

returns the nickname of the default device.

SimpleAiboWebAPI.findDevice — Function

findDevice(;deviceId=nothing, nickname=nothing)

returns a device whose deviceID and/or nickname key(s) match(es) the corresponding parameter(s).

Returns the result of getDefaultDevice() if both parameters are omitted.

Returns Dict() if such a device is absent.

SimpleAiboWebAPI.setDefaultDevice — Function

setDefaultDeviceNickname(;target_deviceId=nothing, target_nickname=nothing)

assigns the default device to the result of findDevice(deviceId=target_deviceId, nickname=target_nickname).

The default device remains unchanged if such a device is absent.

Action APIs

SimpleAiboWebAPI.askActionSimple — Function

askActionSimple(
  api_name,
  arguments,
  target_deviceId,
  timeoutLimit = 10)

asks a device to perform an action, and returns the result by Dict().

api_name and arguments specify the action and arguments.

target_deviceId specifies the deviceId of the target device.

If timeoutLimit <= 0, immediately returns executionId.

Otherwise, invokes getExecution(executionId, timeoutLimit).

timeoutLimit specifies the limit of timeouts in seconds, and defaults to 10 (seconds).

SimpleAiboWebAPI.askAction — Function

askAction(
  api_name,
  arguments = Dict(); 
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks a device to perform an action, and returns the result by Dict().

api_name and arguments specify the action and arguments.

arguments is optional, defaults to Dict().

target device to which the Web API directs are determined as follows using two optional keyword arguments, target_deviceId and target_nickname

the result of getDefaultDevice() if both parameters are omitted.

the result of findDevice(deviceId=target_deviceID, nickname=target_nickname) if either or both parameters are present.

If timeoutLimit <= 0, immediately returns executionId.

Otherwise, invokes getExecution(executionId, timeoutLimit).

timeoutLimit specifies the limit of timeouts in seconds, and defauls to 10 (seconds).

SimpleAiboWebAPI.getExecution — Function

getExecution(
  executionId, 
  timeoutLimit=10 )

receives the result of executionId.

timeoutLimit specifies the limit of timeouts in seconds, and defaults to 10 (seconds).

Advanced APIs

All the advanced API contain "api_name" inside.

All the advanced API have following three optional keyword arguments, as askAction has.

target_deviceID

target_nickname

Above two arguments are hints to determine target device.

timeoutLimit

Status related APIs

SimpleAiboWebAPI.askHungryStatus — Function

askHungryStatus(;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks hungry_status.

This method is equivalent to askAction("hungry_status")

SimpleAiboWebAPI.askSleepyStatus — Function

askSleepyStatus(;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks sleepy_status.

This method is equivalent to askAction("sleepy_status")

SimpleAiboWebAPI.askBodyTouchedStatus — Function

askBodyTouchedStatus(;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks body_touched_status.

This method is equivalent to askAction("body_touched_status")

SimpleAiboWebAPI.askPawPadsStatus — Function

askPawPadsStatus(;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks paw_pads_status.

This method is equivalent to askAction("paw_pads_status")

Posture related APIs

SimpleAiboWebAPI.askPostureStatus — Function

askPostureStatus(;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks posture_status.

This method is equivalent to askAction("posture_status")

SimpleAiboWebAPI.askChangePosture — Function

askChangePosture(
  finalPosture,
  enquee=false;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks change_posture.

finalPosture should be one of the following string:

back

crouch

down

down_and_lengthen_behind

down_and_shorten_behind

sit_and_raise__both_hands

sit

sleep

stand

stand_straight

This method is equivalent to askAction("change_posture", Dict(FinalPosture=>finalPosture, Enqueue=>enqueue))

SimpleAiboWebAPI.askMoveHead — Function

askMoveHead(
  azimuth,
  elevation,
  velocity,
  enquee=false;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks move_head.

azimuth : -80 – 80 (degree)

elevation : -40 – 40 (degree)

velocity : 10 – 80 (deg/sec)

This method is equivalent to askAction("move_head", Dict(Duration=>duration, Enqueue=>enqueue))

SimpleAiboWebAPI.askStay — Function

askStay(
  duration,
  enquee=false;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks stay for duration seconds.

duration : 0–360 (seconds)

This method is equivalent to askAction("stay", Dict(Duration=>duration, Enqueue=>enqueue))

Play related APIs

SimpleAiboWebAPI.askBitingStatus — Function

askBitingStatus(;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks biting_status.

This method is equivalent to askAction("biting_status")

SimpleAiboWebAPI.askPlayBone — Function

askPlayBone(
  enquee=false;
  category="holdMouth",
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks play_bone.

This method is equivalent to askAction("play_bone", Dict(Category=>holdMouth, Enqueue=>enqueue))

SimpleAiboWebAPI.askPlayDice — Function

askPlayDice(
  category="holdMouthDice",
  enquee=false;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks change_posture.

category should be one of the following string:

holdMouthDice

rollDiceLeft

rollDiceRight

rollDicePush

rollDicePull

stackDice

This method is equivalent to askAction("play_dice", Dict(Category=>category, Enqueue=>enqueue))

SimpleAiboWebAPI.askPlayMotion — Function

askPlayMotion(
  category="agree",
  enquee=false;
  mode="NONE",
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks play_motion.

category should be one of the following string. Optional Mode parameters other than "NONE" is also listed.

agree

bad

bark

beckon

Mode: BODY_LEFT

Mode: BODY_RIGHT

belch

bentBack

breath

curious

dance

drawInOnesChin

dreaming

friendly

friendlyPolite

gasp

halfAsleep

handsUp

happyOrHot

heading

Mode: NONE

Mode: SPACE_LEFT

Mode: SPACE_RIGHT

helloIloveYou

highFive

Mode: NONE

Mode: BODY_LEFT

Mode: BODY_RIGHT

imFriendly

jiggle

kiss

lickBody

Mode: BODY_LEFT

Mode: BODY_RIGHT

lookAroundHead

lookOver

marking

Mode: BOY

Mode: GIRL

nodHead

openMouth

overJoyed

paw

Mode: BODY_LEFT

Mode: BODY_RIGHT

peace

Mode: SPACE_LEFT

Mode: SPACE_RIGHT

perceive

playBiting

prettyPlease

ready

relax

restless

rubBack

scratchFloor

scratchHead

Mode: HIND_LEFT

Mode: HIND_RIGHT

shakeHead

shakeHipsBehind

shudder

sideKick

Mode: FRONT_LEFT

Mode: FRONT_RIGHT

sideUp

Mode: BODY_LEFT

Mode: BODY_RIGHT

sing

sleepTalking

sneeze

sniff

sniffDown

sniffUp

startled

startledLittle

stretch

swing

touched

Mode: SPACE_CENTER

waiting

washFace

whine

wiggleEar

Mode : BODY_BOTH

woof

yap

yawn

This method is equivalent to askAction("play_motion", Dict(Category=>category, Mode=>mode, Enqueue=>enqueue))

Voice related APIs

SimpleAiboWebAPI.askNameCalledStatus — Function

askNameCalledStatus(;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks name_called_status.

This method is equivalent to askAction("name_called_status")

SimpleAiboWebAPI.askVoiceCommandStatus — Function

askVoiceCommandStatus(;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks voice_command_status.

This method is equivalent to askAction("voice_command_status")

Person related APIs

SimpleAiboWebAPI.askApproachPerson — Function

askApproachPerson(
  distanceFromTarget=1,
  enquee=false;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks approach_person.

distanceFromTarget : 0.4 – 2 (meter)

This method is equivalent to askAction("approach_person", Dict(DistanceFromTarget=>distanceFromTarget, Enqueue=>enqueue))

SimpleAiboWebAPI.askChasePerson — Function

askChasePerson(
  chasingDurationMsec=15*1000,
  enquee=false;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks chase_person.

chasingDurationMsec : 0 – 360000 (milliseconds)

This method is equivalent to askAction("chase_person", Dict(ChasingDurationMsec=>chasingDurationMsec, Enqueue=>enqueue))

SimpleAiboWebAPI.askFindPerson — Function

askFindPerson(;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks find_person.

This method is equivalent to askAction("find_person")

Position related APIs

SimpleAiboWebAPI.askExplore — Function

askExplore(
  duration=60,
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks explore.

duration : 0 – 360 (seconds)

This method is equivalent to askAction("explore", Dict(Duration=>duration, Enqueue=>enqueue))

SimpleAiboWebAPI.askMoveToPosition — Function

askMoveToPosition(
  targetType="charging_station",
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks move_to_position.

targetType : one of the following constants:

charging_station

greeting_spot

toilet

This method is equivalent to askAction("move_to_position", Dict(TargetType=>targetType, Enqueue=>enqueue))

Moving related APIs

SimpleAiboWebAPI.askTurnAround — Function

askTurnAround(
  turnSpeed=1,
  turnAngle=0,
  enqueue=false;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks turn_around.

turnSpeed : one of 0 (slow), 1, 2 (fast)

turnAngle : -180 – 180, counterclockwise with respect to the current aibo angle.

This method is equivalent to askAction("turn_around", Dict(TurnSpeed=> turnSpeed, TurnAngle=> turnAngle, Enqueue=>enqueue))

SimpleAiboWebAPI.askMoveAlongCircle — Function

askMoveAlongCircle(
  walkSpeed=1,
  radius=1,
  movingAngle=360,
  enqueue=false;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks move_along_circle.

walkSpeed : one of 0 (slow), 1, 2 (fast)

radius : 0.5 – 3 (meter)

movingAngle : 0 – 1080 (degree)

This method is equivalent to askAction("move_along_circle", Dict(WalkSpeed=> walkSpeed, Radius=> radius, MovingAngle=>movingAngle, Enqueue=>enqueue))

SimpleAiboWebAPI.askMoveDirection — Function

askMoveDirection(
  walkSpeed=1,
  targetDistance=1,
  targetAngle=0,
  enqueue=false;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks move_direction.

walkSpeed : one of 0 (slow), 1, 2 (fast)

targetDistance : 0 – 6 (meter)

targetAngle : -180 – 180, counterclockwise with respect to the current aibo angle.

This method is equivalent to askAction("move_direction", Dict(WalkSpeed=> walkSpeed, TargetDistance=> targetDistance, TargetAngle=> targetAngle, Enqueue=>enqueue))

SimpleAiboWebAPI.askMoveForward — Function

askMoveForward(
  walkSpeed=1,
  walkDistance=1,
  enqueue=false;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks move_forward.

walkSpeed : one of 0 (slow), 1, 2 (fast)

walkDistance : -6 – 6 (meter)

This method is equivalent to askAction("move_forward", Dict(WalkSpeed=> walkSpeed, WalkDistance=> walkDistance, Enqueue=>enqueue))

SimpleAiboWebAPI.askMoveSideways — Function

askMoveSideways(
  walkSpeed=1,
  walkDistance=0.5,
  enqueue=false;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks move_sideways.

walkSpeed : one of 0 (slow), 1, 2 (fast)

walkDistance : -6 – 6 (meter)

This method is equivalent to askAction("move_sideways", Dict(WalkSpeed=> walkSpeed, WalkDistance=> walkDistance, Enqueue=>enqueue))

Object related APIs

SimpleAiboWebAPI.askFindObject — Function

askFindObject(
  targetType="pinkball",
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks find_object.

targetType : one of the following constants:

aibo

aibone

dice

pinkball

This method is equivalent to askAction("find_object", Dict(TargetType=>targetType, Enqueue=>enqueue))

SimpleAiboWebAPI.askFoundObjectsStatus — Function

askFoundObjectsStatus(;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks found_objects_status.

This method is equivalent to askAction("found_objects_status")

SimpleAiboWebAPI.askChaseObject — Function

askChaseObject(
  targetType="pinkball",
  chasingDurationMsec=30*1000,
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks chase_object.

targetType : one of the following constants:

aibo

aibone

dice

pinkball

chasingDurationMsec : 0 – 360000 (milliseconds)

This method is equivalent to askAction("chase_object", Dict(TargetType=>targetType, ChasingDurationMsec=> chasingDurationMsec, Enqueue=>enqueue))

SimpleAiboWebAPI.askApproachObject — Function

askApproachObject(
  targetType="pinkball",
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks approach_object.

targetType : one of the following constants:

aibo

aibone

dice

pinkball

This method is equivalent to askAction("approach_object", Dict(TargetType=>targetType, Enqueue=>enqueue))

SimpleAiboWebAPI.askGetCloseToObject — Function

askGetCloseToObject(
  targetType="pinkball",
  distance=0.2,
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks get_close_to_object. Invoke this method after askApproachObject.

targetType : one of the following constants:

aibo

aibone

dice

pinkball

distance : 0.1 – 0.3 (meter)

This method is equivalent to askAction("get_close_to_object", Dict(TargetType=>targetType, Distance=>distance, Enqueue=>enqueue))

SimpleAiboWebAPI.askKickObject — Function

askKickObject(
  targetType="pinkball",
  kickMotion="kick",
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks kick_object.

targetType : one of the following constants:

pinkball

kickMotion : one of the following constants:

kick

heading

This method is equivalent to askAction("kick_object", Dict(TargetType=>targetType, KickMotion=>kickMotion, Enqueue=>enqueue))

SimpleAiboWebAPI.askReleaseObject — Function

askReleaseObject(
  targetType="aibone",
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks release_object.

targetType : one of the following constants:

aibone

dice

This method is equivalent to askAction("release_object", Dict(TargetType=>targetType, Enqueue=>enqueue))

Alphabetical Index

SimpleAiboWebAPI.askAction

SimpleAiboWebAPI.askActionSimple

SimpleAiboWebAPI.askApproachObject

SimpleAiboWebAPI.askApproachPerson

SimpleAiboWebAPI.askBitingStatus

SimpleAiboWebAPI.askBodyTouchedStatus

SimpleAiboWebAPI.askChangePosture

SimpleAiboWebAPI.askChaseObject

SimpleAiboWebAPI.askChasePerson

SimpleAiboWebAPI.askExplore

SimpleAiboWebAPI.askFindObject

SimpleAiboWebAPI.askFindPerson

SimpleAiboWebAPI.askFoundObjectsStatus

SimpleAiboWebAPI.askGetCloseToObject

SimpleAiboWebAPI.askHungryStatus

SimpleAiboWebAPI.askKickObject

SimpleAiboWebAPI.askMoveAlongCircle

SimpleAiboWebAPI.askMoveDirection

SimpleAiboWebAPI.askMoveForward

SimpleAiboWebAPI.askMoveHead

SimpleAiboWebAPI.askMoveSideways

SimpleAiboWebAPI.askMoveToPosition

SimpleAiboWebAPI.askNameCalledStatus

SimpleAiboWebAPI.askPawPadsStatus

SimpleAiboWebAPI.askPlayBone

SimpleAiboWebAPI.askPlayDice

SimpleAiboWebAPI.askPlayMotion

SimpleAiboWebAPI.askPostureStatus

SimpleAiboWebAPI.askReleaseObject

SimpleAiboWebAPI.askSleepyStatus

SimpleAiboWebAPI.askStay

SimpleAiboWebAPI.askTurnAround

SimpleAiboWebAPI.askVoiceCommandStatus

SimpleAiboWebAPI.findDevice

SimpleAiboWebAPI.getAccessToken

SimpleAiboWebAPI.getCurrentDevices

SimpleAiboWebAPI.getDefaultDevice

SimpleAiboWebAPI.getDefaultDeviceId

SimpleAiboWebAPI.getDefaultDeviceNickname

SimpleAiboWebAPI.getDevices

SimpleAiboWebAPI.getExecution

SimpleAiboWebAPI.setDefaultDevice