List of APIs defined in SimpleAiboWebAPI.

Setup APIs

SimpleAiboWebAPI.getDeviceList — Function

getDeviceList(accessToken)

tries to get device list associated to accessToken.

If successful,

returns device list of type Array{Dict}.

assigns accessToken to currentAccessToken, the local variable.

if non-empty device list is found, updates currentDeviceList and defaultDevice. Namely,

assigns found device list to currentDeviceList, the local variable.

assigns the first element of currentDeviceList to defaultDevice, the local variable, if currentDeviceList is not empty.

SimpleAiboWebAPI.getAccessToken — Function

getAccessToken()

returns the current access token, i.e., currentAccessToken, the local variable of this package.

Basic APIs

SimpleAiboWebAPI.getCurrentDeviceList — Function

getCurrentDeviceList()

returns the current device list, i.e., getCurrentDeviceList, the local variable of this package.

SimpleAiboWebAPI.getDefaultDevice — Function

getDefaultDevice()

returns the default device to which the WebAPI directs when target device is not specified in the call of askAction function.

SimpleAiboWebAPI.getDefaultDeviceId — Function

getDefaultDeviceId()

returns the deviceId of the default device.

SimpleAiboWebAPI.getDefaultDeviceNickname — Function

getDefaultDeviceNickname()

returns the nickname of the default device.

SimpleAiboWebAPI.findDevice — Function

findDevice(;deviceId=nothing, nickname=nothing)

returns a device whose deviceID and/or nickname key(s) match(es) the corresponding parameter(s).

Returns the result of getDefaultDevice() if both parameters are omitted.

Returns Dict() if such a device is absent.

SimpleAiboWebAPI.setDefaultDevice — Function

setDefaultDeviceNickname(;target_deviceId=nothing, target_nickname=nothing)

assigns the default device to the result of findDevice(deviceId=target_deviceId, nickname=target_nickname).

The default device remains unchanged if such a device is absent.

Action APIs

SimpleAiboWebAPI.askActionSimple — Function

askActionSimple(
  api_name,
  arguments,
  target_deviceId,
  timeoutLimit = 10)

asks a device to perform an action, and returns the result by Dict().

api_name and arguments specify the action and arguments.

target_deviceId specifies the deviceId of the target device.

If timeoutLimit <= 0, immediately returns executionId.

Otherwise, invokes getExecution(executionId, timeoutLimit).

timeoutLimit specifies the limit of timeouts in seconds, and defaults to 10 (seconds).

SimpleAiboWebAPI.askAction — Function

askAction(
  api_name,
  arguments = Dict(); 
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks a device to perform an action, and returns the result by Dict().

api_name and arguments specify the action and arguments.

arguments is optional, defaults to Dict().

target device to which the Web API directs are determined as follows using two optional keyword arguments, target_deviceId and target_nickname

the result of getDefaultDevice() if both parameters are omitted.

the result of findDevice(deviceId=target_deviceID, nickname=target_nickname) if either or both parameters are present.

If timeoutLimit <= 0, immediately returns executionId.

Otherwise, invokes getExecution(executionId, timeoutLimit).

timeoutLimit specifies the limit of timeouts in seconds, and defauls to 10 (seconds).

SimpleAiboWebAPI.getExecution — Function

getExecution(
  executionId, 
  timeoutLimit=10 )

receives the result of executionId.

timeoutLimit specifies the limit of timeouts in seconds, and defaults to 10 (seconds).

Advanced APIs

All the advanced API contain "api_name" inside.

All the advanced API have following three optional keyword arguments, as askAction has.

target_deviceID

target_nickname

Above two arguments are hints to determine target device.

timeoutLimit

SimpleAiboWebAPI.askHungryStatus — Function

askHungryStatus(;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks hungry_status.

This method is equivalent to askAction("hungry_status")

SimpleAiboWebAPI.askSleepyStatus — Function

askSleepyStatus(;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks sleepy_status.

This method is equivalent to askAction("sleepy_status")

SimpleAiboWebAPI.askPostureStatus — Function

askPostureStatus(;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks posture_status.

This method is equivalent to askAction("posture_status")

SimpleAiboWebAPI.askChangePosture — Function

askChangePosture(
  finalPosture,
  enquee=false;
  target_deviceID=nothing,
  target_nickname=nothing,
  timeoutLimit=10)

asks change_posture.

finalPosture should be one of the following string:

back

crouch

down

down_and_lengthen_behind

down_and_shorten_behind

sit_and_raise__both_hands

sit

sleep

stand

stand_straight

This method is equivalent to askAction("change_posture", Dict(FinalPosture=>finalPosture, Enqueue=>enqueue))

Alphabetical Index

SimpleAiboWebAPI.askAction

SimpleAiboWebAPI.askActionSimple

SimpleAiboWebAPI.askChangePosture

SimpleAiboWebAPI.askHungryStatus

SimpleAiboWebAPI.askPostureStatus

SimpleAiboWebAPI.askSleepyStatus

SimpleAiboWebAPI.findDevice

SimpleAiboWebAPI.getAccessToken

SimpleAiboWebAPI.getCurrentDeviceList

SimpleAiboWebAPI.getDefaultDevice

SimpleAiboWebAPI.getDefaultDeviceId

SimpleAiboWebAPI.getDefaultDeviceNickname

SimpleAiboWebAPI.getDeviceList

SimpleAiboWebAPI.getExecution

SimpleAiboWebAPI.setDefaultDevice